diabetes_companion/requirements.txt
flask==3.1.1
python-dotenv==1.1.0
langchain==0.3.25
langchain-huggingface==0.2.0
langchain-community==0.3.24
langchain-groq==0.3.2
transformers==4.43.0
torch==2.1.0
scikit-learn==1.6.1
numpy==1.26.4
sentence-transformers==2.6.1
faiss-cpu==1.11.0
langdetect==1.0.9
deep-translator==1.11.4
gunicorn==21.2.0